pyproject.toml
[project]
name = "starfleet-command-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "tavily-python>=0.7.12",
]
